feat: image정보, session, 저장구현 course조회시 Sessions 함꼐 조회

+++ src/main/java/nextstep/courses/domain/CourseRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain;
 
 public interface CourseRepository {
-    int save(Course course);
+    Long save(Course course);
 
     Course findById(Long id);
 }
